docs: add a proof of concept storybook rendering the existing documentation.
affects: @fremtind/jkl-accordion-react

Add storybook in order to render components with different states as part of the development
process. Additionally, with some extra plugins for Storybook, it will be able to run regression
tests that will create snapshots of the html and screenshots of the rendered result. This will make
it easier to avoid creating unintentional changes to how the components are rendered.