feat(feedback): rewrote the component to support unmount
affects: @fremtind/jkl-feedback-react

rewriting the component to be a class component, allows us to use the
componentWillUnmount life cycle method. This gives proper support for
running functions on unmount, thus ensuring that the onSubmit function
is run when the component is unmounted in for instance a route change.

BREAKING CHANGE:
rewritten to class component