fix: default variant should be small
This commit removes default variant value from components
which uses the Label component. This allows the default to be
dictated by Label.
This commit also changes the default variant for BaseAutosuggest
from medium to small to match the default value of Label.
The exception is Feedback which still defaults to a medium label.