fix: handle Files with path set better
First make sure that the layout remains correct when a File has the
path prop set. Secondly, change the DOM structure so that the close
icon that is present when onRemove prop is set can be interacted with.
This means that now only the filename itself becomes a click target
when the file has a path to avoid having nested interactive elements.

ISSUES CLOSED: #3891