fix: the \`label\` prop on \`addOnQuestion\` was removed
addOnQuestion.label was never actually used and always overriden
by one of the required labels on the required options[] prop. To
preserve as much as possible of the default behaviour, it is now
needed to sett addOnQuestion to something truthy in order to get
the question rather than providing an unused label. Presets have
been updated to reflect this so there should be no change for
consumers using those.

BREAKING CHANGE:
TS types have changed