fix: copies .d.ts files to where we expect them to be
The dts plugin has for unknown reasons suddenly decided to
dump the generated files in a different subtree. This happens
even with the exact same configuration and the exact same
resolved version from npm so the only remaining explanation
I can think of is that the amount of generated files prompted
the plugin to change the output path. I can find no obvious
way of telling it to do otherwise and I don't like this
structure so for the time being this is a workaround