fix: allows adding the same file twice
Fixes an issue where trying to add the same file twice
would not work, even if it was removed between attemps.
This was caused by the base <input> element being used
as an uncontrolled input. The value of an input of type
file is by default the path of the most recently added
file, so trying to add it again didn't trigger a change
event. This fix makes the input controlled.

ISSUES CLOSED: #4471