fix: tabbing to radio/checkbox now scrolls it into view
When using tab to navigate to a radio/checkbox that is below
the viewport most browsers will scroll untill the input-control
itself is in view and call it a day. In this case, that means
it makes more sense to keep the hidden input control under
the visible label instead of above it.

ISSUES CLOSED: #4649